refactor: RestaurantPhone VO record로 전환
